requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
python-dotenv==1.0.1
google-generativeai==0.3.2
langchain==0.1.9
langchain-google-genai==0.0.11
pillow==10.2.0
python-jose==3.3.0
pydantic==2.6.1
weasyprint==60.2
markdown==3.5.2
markdown2pdf==0.1.0 